Add namespace filtering for Sockend Subscribers.
Sockend instantiates a single Subscriber for a single Publisher for a given namespace, and won't instantiate a second one even if a new Publisher gets up with the same namespace.

This also means that, Publishers that publish different messages on the same namespace cannot be used; the second one to get up will be ignored.