Handle address in use errors in components
When multiple components that want to listen on a free port got
the ports out of portfinder, there could occur a race condition.
This commit makes sure that if an address in use error occurs,
it restarts the process of finding a new port, thereby saving
the daemon from crashing.