Reorganize requester & responder to remove 'ready' event dependency (also fixes #24)
Axon sockets can bind after the socket is created, while the internal
queue starts working. By making use of the internal queue, we make sure
calls to `send` are queued properly and fulfilled after the requester
connects to a responder.

Also, one socket is enough to connect to multiple servers, there is no need
to create new sockets for each server, which also leads to manual
management of leftover sockets. This both greatly simplifies the logic
and improves memory efficiency. So now each component has a single socket.

Also, previously the requester was the server and the responder was the
client. This was counter-intuitive, so now the requester became the client
and the responder became the server.